Address code scanning alerts

+++ engine/src/gfx/star.cpp
@@ -616,7 +616,7 @@ void Stars::ResetPosition(const QVector &cent) {
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 3; j++) {
             for (int k = 0; k < 3; k++) {
-                pos[i * 9 + j * 3 + k].Set((i - 1) * spread, (j - 1) * spread, (k - 1) * spread);
+                pos[i * 9 + j * 3 + k].Set(static_cast<double>(i - 1) * spread, static_cast<double>(j - 1) * spread, static_cast<double>(k - 1) * spread);
                 pos[i * 9 + j * 3 + k] += cent;
             }
         }
@@ -836,4 +836,3 @@ SpriteStarVlist::~SpriteStarVlist() {
         }
     }
 }
-
