+++ engine/src/gfx/vsimage.cpp
@@ -959,14 +959,14 @@ VSError VSImage::WritePNG(unsigned char *data) {
         png_set_swap(png_ptr);
     }
 #endif
-    int stride = (this->img_depth / 8) * (this->img_alpha ? 4 : 3);
+    intmax_t stride = (this->img_depth / 8) * (this->img_alpha ? 4 : 3);
     png_byte **row_pointers = new png_byte *[this->sizeY];
     if (this->flip) {
-        for (int i = this->sizeY - 1, j = 0; i >= 0; i--, ++j) {
+        for (intmax_t i = this->sizeY - 1, j = 0; i >= 0; i--, ++j) {
             row_pointers[j] = (png_byte *) &data[stride * i * sizeX];
         }
     } else {
-        for (unsigned int i = 0; i < this->sizeY; i++) {
+        for (uintmax_t i = 0; i < this->sizeY; i++) {
             row_pointers[i] = (png_byte *) &data[stride * i * sizeX];
         }
     }
