+++ engine/src/gfx/warptrail.cpp
@@ -96,7 +96,7 @@ struct WarpTrail {
         Vector p, q, r;
         r = (end - start).Cast();
         r.Normalize();
-        static float stretch = XMLSupport::parse_float(vs_config->getVariable("graphics", "warp_trail_stretch", "300"));
+        static double stretch = XMLSupport::parse_float(vs_config->getVariable("graphics", "warp_trail_stretch", "300"));
         r *= un->rSize() * stretch;
         q = Vector(0, 1, 0);
         p = r.Cross(q);
@@ -153,4 +153,3 @@ Mesh *GetWarpMesh(int faction, warptrails *w) {
     }
     return w->factions[faction];
 }
-
