+++ engine/src/physics.cpp
@@ -105,7 +105,7 @@ void PhysicsSystem::Rotate(const Vector &axis) {
 }
 
 void PhysicsSystem::JettisonReactionMass(const Vector &Direction, float speed, float mass) {
-    NetForce += Direction * (speed * mass / GetElapsedTime());
+    NetForce += Direction * (static_cast<double>(speed) * static_cast<double>(mass) / GetElapsedTime());
 }
 
 void PhysicsSystem::JettisonMass(const Vector &Direction, float speed, float jmass) {
@@ -188,4 +188,3 @@ void PhysicsSystem::ApplyImpulses(float Time) {
     *pos += (Velocity + .5 * tempforce).Cast();
     Velocity += tempforce;
 }
-
